Better support files with multiple licenses.
In cases where two "exact match" license files run together, they will be
found by exact match regex -- but the tokenization across the license
boundaries won't match, and can produce a panic when attempting to extract
the text (because the "final token" of the known license text can never be
found in the input, as it has been normalized away: so "or
litigation.Copyright 2012" becomes "or ligitigationcopyright 2012" when
tokenized). If we pass the part of the file that is intended to match the
final token, use that token even if it isn't an exact match.